Split AccountManager into interface and implementation
